server: move the decommission logic out of server.go
Release note: None
