+++ pkg/server/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "config.go",
         "config_unix.go",
         "config_windows.go",
+        "decommission.go",
         "doc.go",
         "drain.go",
         "env_sampler.go",
