+++ pkg/server/server.go
@@ -18,7 +18,6 @@ import (
 	"net/http"
 	"path/filepath"
 	"reflect"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -74,7 +73,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/httputil"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
-	"github.com/cockroachdb/cockroach/pkg/util/log/eventpb"
 	"github.com/cockroachdb/cockroach/pkg/util/log/logcrash"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
@@ -89,7 +87,6 @@ import (
 	"github.com/cockroachdb/sentry-go"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
-	grpcstatus "google.golang.org/grpc/status"
 )
 
 // Server is the cockroach server node.
@@ -211,24 +208,7 @@ func NewServer(cfg Config, stopper *stop.Stopper) (*Server, error) {
 	}
 	stopper.AddCloser(&engines)
 
-	nodeTombStorage := &nodeTombstoneStorage{engs: engines}
-	checkPingFor := func(ctx context.Context, nodeID roachpb.NodeID, errorCode codes.Code) error {
-		ts, err := nodeTombStorage.IsDecommissioned(ctx, nodeID)
-		if err != nil {
-			// An error here means something very basic is not working. Better to terminate
-			// than to limp along.
-			log.Fatalf(ctx, "unable to read decommissioned status for n%d: %v", nodeID, err)
-		}
-		if !ts.IsZero() {
-			// The node was decommissioned.
-			return grpcstatus.Errorf(errorCode,
-				"n%d was permanently removed from the cluster at %s; it is not allowed to rejoin the cluster",
-				nodeID, ts,
-			)
-		}
-		// The common case - target node is not decommissioned.
-		return nil
-	}
+	nodeTombStorage, checkPingFor := getPingCheckDecommissionFn(engines)
 
 	rpcCtxOpts := rpc.ContextOptions{
 		TenantID:  roachpb.SystemTenantID,
@@ -1634,91 +1614,6 @@ func (s *Server) startServeUI(
 	return s.stopper.RunAsyncTask(workersCtx, "server-http", func(context.Context) {
 		netutil.FatalIfUnexpected(connManager.Serve(httpLn))
 	})
-}
-
-// Decommission idempotently sets the decommissioning flag for specified nodes.
-func (s *Server) Decommission(
-	ctx context.Context, targetStatus livenesspb.MembershipStatus, nodeIDs []roachpb.NodeID,
-) error {
-	// If we're asked to decommission ourself we may lose access to cluster RPC,
-	// so we decommission ourself last. We copy the slice to avoid mutating the
-	// input slice.
-	if targetStatus == livenesspb.MembershipStatus_DECOMMISSIONED {
-		orderedNodeIDs := make([]roachpb.NodeID, len(nodeIDs))
-		copy(orderedNodeIDs, nodeIDs)
-		sort.SliceStable(orderedNodeIDs, func(i, j int) bool {
-			return orderedNodeIDs[j] == s.NodeID()
-		})
-		nodeIDs = orderedNodeIDs
-	}
-
-	var event eventpb.EventPayload
-	var nodeDetails *eventpb.CommonNodeDecommissionDetails
-	if targetStatus.Decommissioning() {
-		ev := &eventpb.NodeDecommissioning{}
-		nodeDetails = &ev.CommonNodeDecommissionDetails
-		event = ev
-	} else if targetStatus.Decommissioned() {
-		ev := &eventpb.NodeDecommissioned{}
-		nodeDetails = &ev.CommonNodeDecommissionDetails
-		event = ev
-	} else if targetStatus.Active() {
-		ev := &eventpb.NodeRecommissioned{}
-		nodeDetails = &ev.CommonNodeDecommissionDetails
-		event = ev
-	} else {
-		panic("unexpected target membership status")
-	}
-	event.CommonDetails().Timestamp = timeutil.Now().UnixNano()
-	nodeDetails.RequestingNodeID = int32(s.NodeID())
-
-	for _, nodeID := range nodeIDs {
-		statusChanged, err := s.nodeLiveness.SetMembershipStatus(ctx, nodeID, targetStatus)
-		if err != nil {
-			if errors.Is(err, liveness.ErrMissingRecord) {
-				return grpcstatus.Error(codes.NotFound, liveness.ErrMissingRecord.Error())
-			}
-			return err
-		}
-		if statusChanged {
-			nodeDetails.TargetNodeID = int32(nodeID)
-			// Ensure an entry is produced in the external log in all cases.
-			log.StructuredEvent(ctx, event)
-
-			// If we die right now or if this transaction fails to commit, the
-			// membership event will not be recorded to the event log. While we
-			// could insert the event record in the same transaction as the liveness
-			// update, this would force a 2PC and potentially leave write intents in
-			// the node liveness range. Better to make the event logging best effort
-			// than to slow down future node liveness transactions.
-			if err := s.db.Txn(ctx, func(ctx context.Context, txn *kv.Txn) error {
-				return sql.InsertEventRecord(
-					ctx,
-					s.sqlServer.execCfg.InternalExecutor,
-					txn,
-					int32(s.NodeID()), /* reporting ID: the node where the event is logged */
-					sql.LogToSystemTable|sql.LogToDevChannelIfVerbose, /* we already call log.StructuredEvent above */
-					int32(nodeID), /* target ID: the node that we wee a membership change for */
-					event,
-				)
-			}); err != nil {
-				log.Ops.Errorf(ctx, "unable to record event: %+v: %+v", event, err)
-			}
-		}
-
-		// Similarly to the log event above, we may not be able to clean up the
-		// status entry if we crash or fail -- the status entry is inline, and
-		// thus cannot be transactional. However, since decommissioning is
-		// idempotent, we can attempt to remove the key regardless of whether
-		// the status changed, such that a stale key can be removed by
-		// decommissioning the node again.
-		if targetStatus.Decommissioned() {
-			if err := s.db.PutInline(ctx, keys.NodeStatusKey(nodeID), nil); err != nil {
-				log.Errorf(ctx, "unable to clean up node status data for node %d: %s", nodeID, err)
-			}
-		}
-	}
-	return nil
 }
 
 // Stop stops the server.
